requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.1
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
multidict==6.6.3
openai==1.95.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
